renamed to functional test
