+++ jest.func.config.js
@@ -1,4 +1,4 @@
 module.exports = {
   testEnvironment: 'node',
-  testRegex: './e2e/e2e.js'
+  testRegex: 'func/(.|/)*.test.js'
 }
