+++ package.json
@@ -8,6 +8,16 @@
     "node": ">=10.0.0"
   },
   "repository": "adobe/aio-tvm",
+  "scripts": {
+    "lint": "eslint './actions/**/*.js' './lib/**/*.js' './test/**/*.js' './scripts/**/*.js'",
+    "beautify": "npm run lint -- --fix",
+    "deploy": "webpack && dotenv cna-scripts deploy.actions",
+    "undeploy": "dotenv cna-scripts undeploy.actions",
+    "test": "npm run unit && npm run lint",
+    "unit": "jest",
+    "func": "jest -c jest.func.config.js"
+  },
+  "jest.pathToConfig": "jest.config.js",
   "devDependencies": {
     "@adobe/io-cna-scripts": "^0.1.0",
     "@babel/plugin-transform-runtime": "^7.6.2",
@@ -30,15 +40,6 @@
     "node-fetch": "^2.6.0",
     "webpack": "^4.41.0",
     "webpack-cli": "^3.3.9"
-  },
-  "scripts": {
-    "lint": "eslint './actions/**/*.js' './lib/**/*.js' './test/**/*.js' './scripts/**/*.js'",
-    "beautify": "npm run lint -- --fix",
-    "deploy": "webpack && dotenv cna-scripts deploy.actions",
-    "undeploy": "dotenv cna-scripts undeploy.actions",
-    "test": "npm run unit && npm run lint",
-    "unit": "jest -c jest.config.js",
-    "e2e": "jest -c jest.e2e.config.js"
   },
   "dependencies": {
     "@adobe/aio-lib-core-logging": "0.0.4",
