+++ test/unit/actions/azure-blob.test.js
@@ -9,10 +9,10 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-const azureBlobAction = require('../../actions/azure-blob')
+const azureBlobAction = require('../../../actions/azure-blob')
 
-const { AzureBlobTvm } = require('../../lib/impl/AzureBlobTvm')
-jest.mock('../../lib/impl/AzureBlobTvm')
+const { AzureBlobTvm } = require('../../../lib/impl/AzureBlobTvm')
+jest.mock('../../../lib/impl/AzureBlobTvm')
 
 beforeEach(() => {
   AzureBlobTvm.prototype.processRequest.mockReset()
