+++ test/unit/actions/azure-cosmos.test.js
@@ -9,10 +9,10 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-const azureCosmosAction = require('../../actions/azure-cosmos')
+const azureCosmosAction = require('../../../actions/azure-cosmos')
 
-const { AzureCosmosTvm } = require('../../lib/impl/AzureCosmosTvm')
-jest.mock('../../lib/impl/AzureCosmosTvm')
+const { AzureCosmosTvm } = require('../../../lib/impl/AzureCosmosTvm')
+jest.mock('../../../lib/impl/AzureCosmosTvm')
 
 beforeEach(() => {
   AzureCosmosTvm.prototype.processRequest.mockReset()
