+++ test/unit/lib/impl/AwsS3Tvm.test.js
@@ -10,7 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-const { AwsS3Tvm } = require('../../../lib/impl/AwsS3Tvm')
+const { AwsS3Tvm } = require('../../../../lib/impl/AwsS3Tvm')
 
 const aws = require('aws-sdk')
 jest.mock('aws-sdk')
